wp-backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
Django==5.2.12
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
msgpack==1.1.2
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
redis==7.3.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3

